updata

+++ DynamicPush/src/main/kotlin/DynamicPushData.kt
@@ -7,23 +7,23 @@ import net.mamoe.mirai.console.data.value
 object DynamicPushData : AutoSavePluginData("DynamicPushData") { // "name" 是保存的文件名 (不带后缀)
 
     @ValueDescription("历史动态时间戳")
-    val timeStampOfDynamic: MutableMap<Int, Long> by value(
+    val timeStampOfDynamic: MutableMap<Long, Long> by value(
         mutableMapOf(
-            233114659 to 1L,
-            161775300 to 1L,
-            233108841 to 1L,
-            401742377 to 1L
+            233114659L to 1L,
+            161775300L to 1L,
+            233108841L to 1L,
+            401742377L to 1L
         )
     )
 
 
     @ValueDescription("订阅清单")
-    val subscriptionList: MutableMap<Int, MutableSet<Long>> by value(
+    val subscriptionList: MutableMap<Long, MutableSet<Long>> by value(
         mutableMapOf(
-            233114659 to mutableSetOf(),
-            161775300 to mutableSetOf(),
-            233108841 to mutableSetOf(),
-            401742377 to mutableSetOf()
+            233114659L to mutableSetOf(),
+            161775300L to mutableSetOf(),
+            233108841L to mutableSetOf(),
+            401742377L to mutableSetOf()
         )
     )
 }
